+++ MetaFabSwift5/Classes/MetaFabSwift5/Models/BatchMintCollectionItemsRequest.swift
@@ -15,13 +15,13 @@ public struct BatchMintCollectionItemsRequest: Codable, JSONEncodable, Hashable 
     /** A valid EVM based address to create (mint) the items for. For example, `0x39cb70F972E0EE920088AeF97Dbe5c6251a9c25D`. */
     public var address: String?
     /** An array of unique itemIds to create (mint). */
-    public var itemIds: [Double]
+    public var itemIds: [Int]
     /** An array of the quantities of each of the unique itemIds provided to create (mint). The quantity of each itemId in itemIds should be at the same index as the specific itemId in the itemIds array. For example, quantities[2] defines the quantity to mint for itemIds[2], etc. */
-    public var quantities: [Double]
+    public var quantities: [Int]
     /** Any wallet id within the MetaFab ecosystem to create (mint) the items for. */
     public var walletId: String?
 
-    public init(address: String? = nil, itemIds: [Double], quantities: [Double], walletId: String? = nil) {
+    public init(address: String? = nil, itemIds: [Int], quantities: [Int], walletId: String? = nil) {
         self.address = address
         self.itemIds = itemIds
         self.quantities = quantities
